Windows build fixes.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@41771 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/VMM/VMMAll/EMAll.cpp
@@ -352,7 +352,7 @@ static DECLCALLBACK(int) emReadBytes(PDISCPUSTATE pDis, uint8_t offInstr, uint8_
         }
     }
 
-    pDis->cbCachedInstr = offInstr + cbToRead;
+    pDis->cbCachedInstr = offInstr + (uint8_t)cbToRead;
     return rc;
 }
